requirements.txt
discord.py
python-dotenv
peewee
requests
xmltodict
python-dateutil
pymysql
peewee-migrations